code reformat, added git-related file, casted some variables

+++ ActiverseEngine/Actor.java
@@ -80,7 +80,7 @@ public abstract class Actor {
 
             AffineTransform old = g2d.getTransform();
 
-            g2d.rotate(direction, x + width / 2, y + height / 2);
+            g2d.rotate(direction, x + (double) width / 2, y + (double) height / 2);
             g2d.drawImage(image.getImage(), x, y, null);
 
             g2d.setTransform(old);
@@ -233,6 +233,7 @@ public abstract class Actor {
     public void setHeight(int height) {
         this.height = height;
     }
+
     /**
      * Sets the x-coordinate of the actor's location.
      *
